+++ src/ts/components/videoqualityselectbox.ts
@@ -58,6 +58,8 @@ export class VideoQualitySelectBox extends SelectBox {
     player.addEventHandler(player.EVENT.ON_SOURCE_UNLOADED, updateVideoQualities);
     // Update qualities when a new source is loaded
     player.addEventHandler(player.EVENT.ON_READY, updateVideoQualities);
+    // Update qualities when the period within a source changes
+    player.addEventHandler(player.EVENT.ON_PERIOD_SWITCHED, updateVideoQualities);
     // Update quality selection when quality is changed (from outside)
     if (player.EVENT.ON_VIDEO_QUALITY_CHANGED) {
       // Since player 7.3.1